refactor(core): move MemoryBackend to core, delete the adapter it forced
`rustykrab-memory` sits below `rustykrab-tools` in the crate graph, and
`MemoryBackend` was declared in `rustykrab-tools`. So the crate that owns
agent memory could not implement the trait describing agent memory. It said
so in a comment, carried a structurally identical but unrelated method set,
and the binary bridged the two with a 44-line adapter whose five methods
were four delegations and one argument parse.

The trait moves to `rustykrab-core`, next to `Tool` and `ModelProvider`.
`HybridMemoryBackend` implements it directly and the adapter is gone. The
argument parse goes with it, which is where it belonged: deciding what an
unparseable conversation id means — fail, or widen and say so — needs to
know what widening costs, and that is the memory crate's knowledge, not the
binary's. `rustykrab_tools::MemoryBackend` still resolves, as a re-export.

The other five backend traits stay where they are. Every one of their real
implementors — `SubagentRunner`, `EnigoXcapBackend`, `CronAdapter`,
`MessageAdapter`, `VideoChannelAdapter` — already lives in `rustykrab-cli`,
`rustykrab-agent` or `rustykrab-tools` itself, all at or above the tool
crate, so none of them is blocked and moving would be churn. `MemoryBackend`
was the only one with a cycle to break, and the only one that had produced
a pass-through adapter.

CLAUDE.md described the pattern as uniform; it now states the rule that
actually applies, and that a pure pass-through adapter is the symptom of a
trait in the wrong crate.

Found by the architecture review; see docs/architecture/02-extension-seams.md.
